Create oled1.ino
Parking aid. How to park your car safely.
Changing the frequency of the LED light and the loud sound as you approach the end of the garage.
Show message's on the OLED display 124x32.
This version have INTRO SONG - STAR WARS SONG
In the future, it is possible to send information via the Internet about the car being left in the garage.